% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseproject-package.R
\docType{package}
\name{courseproject-package}
\alias{courseproject}
\alias{courseproject-package}
\title{courseproject: PPP}
\description{
Build Pretty Plots Fast using ggplot2
}
\author{
\strong{Maintainer}: Laura Iisager \email{leij@clin.au.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
