% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meantable.R
\name{meantable}
\alias{meantable}
\title{meantable()}
\usage{
meantable(x, colvar, rowvar, statvar)
}
\arguments{
\item{x}{}

\item{colvar}{}

\item{rowvar}{}

\item{statvar}{}
}
\value{
a table
}
\description{
meantable()
}
\examples{
meantable(fl, origin, carrier, air_time)

}
