% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediantable.R
\name{mediantable}
\alias{mediantable}
\title{mediantable()}
\usage{
mediantable(x, colvar, rowvar, statvar)
}
\arguments{
\item{x}{}

\item{colvar}{}

\item{rowvar}{}

\item{statvar}{}
}
\description{
mediantable()
}
\examples{
mediantable(fl, origin, carrier, air_time)
}
