% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp2.R
\name{ppp_barplot}
\alias{ppp_barplot}
\title{Build pretty barplots fast}
\usage{
ppp_barplot(data, x_var, colors)
}
\arguments{
\item{data}{a data frame}

\item{x_var}{a categorical variable in data}

\item{colors}{character vector of colors names for x_var}
}
\value{
A pretty colored barplot with counts printed on the plot
}
\description{
A function to produce barplots using ggplot2 with theme, colors, and text
}
\examples{
my_colors <- c("#337495", "#002a33", "#2f5a69", "#143b44")
ppp_barplot(data = dplyr::starwars, x_var = "sex", colors = my_colors)
}
