% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-package.R
\docType{package}
\name{project-package}
\alias{project}
\alias{project-package}
\title{project: Threeway summary tables}
\description{
This package makes it easy to present summary statistics of one variable grouped by two other variables.
}
\author{
\strong{Maintainer}: Gitte S Brix \email{gittbi@rm.dk} (\href{https://orcid.org/0000-0002-7426-2818}{ORCID})

}
\keyword{internal}
