% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseproject-package.R
\docType{package}
\name{courseproject-package}
\alias{courseproject}
\alias{courseproject-package}
\title{courseproject: PPP}
\description{
Build pretty plots fast using ggplot2.
}
\author{
\strong{Maintainer}: Eva Ferlev Jensby \email{evaferlev@clin.au.dk}

}
\keyword{internal}
