% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp2.R
\name{ppp_barplot}
\alias{ppp_barplot}
\title{Build pretty bar plots fast}
\usage{
ppp_barplot(data, x_var, colors)
}
\arguments{
\item{data}{data frame}

\item{x_var}{categorical variable in data}

\item{colors}{character vector of color names of same length as x_var}
}
\value{
pretty bar plot with counts and percentages printed on the plot.

#' @import ggplot2
}
\description{
A Function to produce bar plots using ggplot2 with theme, colors and counts (percentages)
}
\examples{

my_colors <- c('#e09f3e', '#9e2a2b','#540b0e', '#335c67')

ppp_barplot(data = dplyr::starwars,
           x_var = "sex",
           colors = my_colors)


}
