% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp2.R
\name{ppp_barplot}
\alias{ppp_barplot}
\title{Build pretty bar plots fast}
\usage{
ppp_barplot(data, x_var, colors)
}
\arguments{
\item{data}{The data frame containing the variables to be plotted.}

\item{x_var}{The categorical variable in the data frame that will be used to create the bar plot.}

\item{colors}{A character vector of color names, where the number of colors provided must match the number of unique categories in the x_var variable.}
}
\value{
A visually appealing bar plot displaying counts and percentages for each category in x_var, with the values printed directly on the plot.
}
\description{
A Function to produce bar plots using ggplot2 with theme, colors and counts (percentages)
}
\examples{

my_colors <- c('#e09f3e', '#9e2a2b','#540b0e', '#335c67')

ppp_barplot(data = dplyr::starwars,
           x_var = "sex",
           colors = my_colors)


}
