% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp2.R
\name{ppp_barplot}
\alias{ppp_barplot}
\title{Build pretty bar plots fast}
\usage{
ppp_barplot(data, x_var, colors)
}
\arguments{
\item{data}{data frame}

\item{colors}{character vector of color names of same length as x_var}

\item{x_svar}{categorical variable in data}
}
\value{
A pretty colored bar plot with counts and percentages printed on the plot
}
\description{
A function to produce bar plots using ggplot2 with theme, colors, and counts (percentages)
}
\examples{
my_colors <- c("#E09F3E", "#9E2A2B", "#540B0E", "#335C67")
ppp_barplot(data = dplyr::starwars, x_var = "sex", colors = my_colors)
ppp_barplot(data = iris, x_var = "Species", colors = c("#337495", "#002a33", "#2f5a69"))
}
