% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{ppp_boxplot}
\alias{ppp_boxplot}
\title{Build pretty boxplots fast}
\usage{
ppp_boxplot(data, x_var, y_var, comparisons, colors, method)
}
\arguments{
\item{data}{a data frame}

\item{x_var}{categorical variable in data}

\item{y_var}{numeric variable in data}

\item{comparisons}{list of groups to compare}

\item{colors}{character vector of color names for x_var}

\item{method}{statistical method}
}
\value{
A pretty boxplot with stats printed on the plot.
}
\description{
A function to produce boxplots using ggplot2 with theme, colors and statistics.
}
\examples{
my_comparisons <- list(c("female", "hermaphroditic"), c("female", "male"), c("female", "none"))
my_colors <- c('#47B2EC','#162F82', '#608D00', '#005115')
my_method = 'wilcox.test'
ppp_boxplot(data = dplyr::starwars,
           x_var = "sex",
           y_var = "height",
           comparisons = my_comparisons,
           colors = my_colors,
           method = my_method)
}
